test: reuse shared ACP fixture in session tests
Move the remaining session-only fake-agent behaviors into the shared ACP test fixture so workflow and ACP tests exercise the same script.
